Merged PR 193544: signature part of z is encoded and decoded
signature part of z is encoded and decoded.
There is a separated folder for each submodule.

----
#### AI description  (iteration 1)
#### PR Classification
Design and implementation of encoding and decoding submodules for signature part `z`.

#### PR Summary
This pull request introduces the RTL design and testbenches for the `SigEncode` and `SigDecode_z` submodules, which handle encoding and decoding of the coefficients of `z` into and from 20-bit values.
- Added `sigdecode_z_tb.sv` to implement and test the `SigDecode_z` module.
- Added `sigencode_tb.sv` to implement and test the `SigEncode` module.

Related work items: #752340, #752342